+++ src/Molecules/FlexTable/Cell/Cell.tsx
@@ -4,7 +4,7 @@ import * as R from 'ramda';
 import { isElement, isFunction } from '../../../common/utils';
 import { Flexbox } from '../../..';
 import { useColumnLayout } from '../shared/ColumnProvider';
-import { CellComponent } from './Cell.types';
+import { CellComponent, InnerCellComponent } from './Cell.types';
 import { TextWrapper } from './TextWrapper';
 import { useFlexTable } from '../shared/FlexTableProvider';
 
@@ -12,9 +12,18 @@ const StyledFlexbox = styled(Flexbox)`
   overflow: hidden;
 `;
 
-// TODO: Fix typings that react memo causes when exporting textwrapper on cell
-// @ts-ignore
-const Cell: CellComponent = React.memo(({ children, className, columnId }) => {
+const InnerCell: InnerCellComponent = React.memo(
+  ({ children, className, columnId, flexProps, fontSize }) => (
+    <StyledFlexbox className={className} role="cell" {...flexProps}>
+      {isElement(children) && children}
+      {isFunction(children)
+        ? children({ fontSize, columnId })
+        : !isElement(children) && <TextWrapper fontSize={fontSize}>{children}</TextWrapper>}
+    </StyledFlexbox>
+  ),
+);
+
+const Cell: CellComponent = ({ children, className, columnId }) => {
   const [columnLayout] = useColumnLayout(columnId);
   const { fontSize } = useFlexTable();
 
@@ -23,14 +32,16 @@ const Cell: CellComponent = React.memo(({ children, className, columnId }) => {
   }
 
   return (
-    <StyledFlexbox className={className} role="cell" {...columnLayout.flexProps}>
-      {isElement(children) && children}
-      {isFunction(children)
-        ? children({ fontSize, columnId })
-        : !isElement(children) && <TextWrapper fontSize={fontSize}>{children}</TextWrapper>}
-    </StyledFlexbox>
+    <InnerCell
+      className={className}
+      columnId={columnId}
+      flexProps={columnLayout.flexProps}
+      fontSize={fontSize}
+    >
+      {children}
+    </InnerCell>
   );
-});
+};
 
 Cell.TextWrapper = React.memo(TextWrapper);
 export default Cell;