+++ src/Molecules/FlexTable/Cell/Cell.types.ts
@@ -1,5 +1,6 @@
 import React, { ReactNode } from 'react';
 import { FontSize, Density } from '../shared/shared.types';
+import { FlexPropsType } from '../shared/ColumnProvider/ColumnProvider.types';
 
 type RenderPropArguments = { density: Density; fontSize: FontSize; columnId: string };
 type RenderFunc = (props: RenderPropArguments) => ReactNode;
@@ -20,6 +21,8 @@ export type TextWrapperComponent = React.FC<TextWrapperProps>;
 export type CellComponents = { TextWrapper: TextWrapperComponent };
 
 export type CellComponent = React.FC<Props> & CellComponents;
+
+export type InnerCellComponent = React.FC<Props & TextWrapperProps & { flexProps: FlexPropsType }>;
 
 type ExpandCellProps = {
   disabled?: boolean;