+++ src/Molecules/FlexTable/FlexTable.stories.tsx
@@ -204,6 +204,15 @@ export const TableWithDifferentRows = () => (
 );
 
 export const TableWithDifferentColumns = () => {
+  const FlagCell: React.FC<{ columnId: string }> = React.memo(({ children, columnId }) => (
+    <FlexTable.Cell columnId={columnId}>
+      <FlexTable.CellInlineContainer>
+        <FlexTable.Cell.TextWrapper>{children}</FlexTable.Cell.TextWrapper>
+        <Flag country="SE" />
+      </FlexTable.CellInlineContainer>
+    </FlexTable.Cell>
+  ));
+
   const ColumnWidthTableExample = () => (
     <StyledFlexTable>
       <FlexTable.HeaderRow>
@@ -230,7 +239,47 @@ export const TableWithDifferentColumns = () => {
       <FlexTable.Row>
         <FlexTable.Cell columnId="column1">Cell 2-1</FlexTable.Cell>
         <FlexTable.Cell columnId="column2">Cell 2-2</FlexTable.Cell>
-        <FlexTable.Cell columnId="column3">Cell 2-3</FlexTable.Cell>
+        <FlagCell columnId="column3">Long cell with flag</FlagCell>
+        <FlexTable.Cell columnId="column4">Cell 2-4</FlexTable.Cell>
+      </FlexTable.Row>
+      <FlexTable.Row>
+        <FlexTable.Cell columnId="column1">Cell 3-1</FlexTable.Cell>
+        <FlexTable.Cell columnId="column2">Cell 3-2</FlexTable.Cell>
+        <FlexTable.Cell columnId="column3">Cell 3-3</FlexTable.Cell>
+        <FlexTable.Cell columnId="column4">Cell 3-4</FlexTable.Cell>
+      </FlexTable.Row>
+    </StyledFlexTable>
+  );
+
+  const ColumnWidthSortableTableExample = () => (
+    <StyledFlexTable>
+      <FlexTable.HeaderRow>
+        <FlexTable.Header sortable flex="1" columnId="column1">
+          Flex 1
+        </FlexTable.Header>
+        <FlexTable.Header sortable flex="0 15%" columnId="column2">
+          Fifteen percent
+        </FlexTable.Header>
+        <FlexTable.Header sortable flex="0 100px" columnId="column3">
+          Loooooooooooooong header set width
+        </FlexTable.Header>
+        <FlexTable.Header sortable columnId="column4">
+          Default
+        </FlexTable.Header>
+      </FlexTable.HeaderRow>
+      <FlexTable.Row>
+        <FlexTable.Cell columnId="column1">Cell 1-1</FlexTable.Cell>
+        <FlexTable.Cell columnId="column2">
+          Very long cell content that should be truncated or ellipsized depepending on your language
+          preferences
+        </FlexTable.Cell>
+        <FlexTable.Cell columnId="column3">Cell 1-3</FlexTable.Cell>
+        <FlexTable.Cell columnId="column4">Cell 1-4</FlexTable.Cell>
+      </FlexTable.Row>
+      <FlexTable.Row>
+        <FlexTable.Cell columnId="column1">Cell 2-1</FlexTable.Cell>
+        <FlexTable.Cell columnId="column2">Cell 2-2</FlexTable.Cell>
+        <FlagCell columnId="column3">Long cell with flag</FlagCell>
         <FlexTable.Cell columnId="column4">Cell 2-4</FlexTable.Cell>
       </FlexTable.Row>
       <FlexTable.Row>
@@ -310,6 +359,8 @@ export const TableWithDifferentColumns = () => {
       <DifferentAlignmentsTableExample />
       <Typography type="title3">Table With Column Width Set</Typography>
       <ColumnWidthTableExample />
+      <Typography type="title3">Sortable Table With Column Width Set</Typography>
+      <ColumnWidthSortableTableExample />
       <Typography type="title3">Table With Columns Hidden Based On Screen Size</Typography>
       <MediaColumnsExample />
     </StyledDiv>
@@ -318,17 +369,29 @@ export const TableWithDifferentColumns = () => {
 
 export const TableWithDifferentHeaders = () => {
   const DefaultTableHeaders = () => {
+    const StyledFlexboxContainer = styled(Flexbox)`
+      justify-content: inherit;
+    `;
+
     const CustomisedTableHeader: React.FC = ({ children }) => (
       <FlexTable.Header columnId="column3" sortable>
         {({ sortable, sorted, fontSize, onSortClick, sortOrder }) => (
           <FlexTable.Header.SortButton onClick={onSortClick}>
-            <>
-              <Flag country="SE" inline height={3} />
-              <FlexTable.Header.TextWrapper fontSize={fontSize} sorted={sorted}>
-                {children}
-              </FlexTable.Header.TextWrapper>
-              {sortable && <FlexTable.Header.SortIcon sortOrder={sortOrder} />}
-            </>
+            <StyledFlexboxContainer container>
+              <Flexbox item>
+                <Flag country="SE" inline height={3} />
+              </Flexbox>
+              <FlexTable.CellInlineContainer item>
+                <FlexTable.Header.TextWrapper fontSize={fontSize} sorted={sorted}>
+                  {children}
+                </FlexTable.Header.TextWrapper>
+              </FlexTable.CellInlineContainer>
+              {sortable && (
+                <Flexbox item>
+                  <FlexTable.Header.SortIcon sortOrder={sortOrder} />
+                </Flexbox>
+              )}
+            </StyledFlexboxContainer>
           </FlexTable.Header.SortButton>
         )}
       </FlexTable.Header>