+++ src/Molecules/FlexTable/FlexTable.types.ts
@@ -5,6 +5,7 @@ import { FooterComponent } from './Footer/Footer.types';
 import { constants } from './shared';
 import { CellComponent, ExpandCellComponent } from './Cell/Cell.types';
 import { Props as FlexTableProviderProps } from './shared/FlexTableProvider/FlexTableProvider.types';
+import { CellInlineContainerComponent } from './shared/CellInlineContainer/CellInlineContainer.types';
 
 type HtmlProps = {} & React.HTMLAttributes<HTMLDivElement>;
 
@@ -24,5 +25,6 @@ export type FlexTableComponents = {
   Row: RowComponent;
   Cell: CellComponent;
   ExpandCell: ExpandCellComponent;
+  CellInlineContainer: CellInlineContainerComponent;
   CONSTANTS: typeof constants;
 };