+++ src/Molecules/FlexTable/Row/Row.tsx
@@ -44,6 +44,16 @@ const StyledRow = styled(Flexbox).withConfig({
 
   padding-top: ${p => getDensityPaddings(p.density)}px;
   padding-bottom: ${p => getDensityPaddings(p.density)}px;
+
+  & > * {
+    &:first-child {
+      padding-left: 0;
+    }
+
+    &:last-child {
+      padding-right: 0;
+    }
+  }
 `;
 
 const StyledExpandedRow = styled('div').withConfig({
@@ -105,6 +115,7 @@ const Row: RowComponent & RowComponents = ({
             separatorColor={separatorColor}
             density={density}
             expandable={expandable}
+            gutter={1}
             {...htmlProps}
           >
             {component}