+++ src/Molecules/FlexTable/Row/components/ExpandElement.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { ExpandProps } from '../Row.types';
 import FlexTable from '../..';
-import { ICON_COLUMN_DEFAULT_FLEX_PROPS, COLUMN_ID_EXPAND } from '../../shared/constants';
+import { ExpandProps } from '../Row.types';
 import { ExpandCell } from '../../Cell/ExpandCell';
+import { ICON_COLUMN_DEFAULT_FLEX_PROPS, COLUMN_ID_EXPAND } from '../../shared/constants';
 
 export const ExpandElement: React.FC<
   ExpandProps & { isContent: boolean; disabled?: boolean; setExpand: (expanded: boolean) => void }