+++ src/Molecules/FlexTable/shared/index.ts
@@ -1,5 +1,14 @@
+import { CellInlineContainer } from './CellInlineContainer';
 import { ColumnProvider } from './ColumnProvider';
 import { RenderForSizes, getScreenMedia } from './RenderForSizes';
 import * as constants from './constants';
+import { StyledTruncateTooltip } from './StyledTruncateTooltip';
 
-export { constants, ColumnProvider, RenderForSizes, getScreenMedia };
+export {
+  constants,
+  ColumnProvider,
+  RenderForSizes,
+  getScreenMedia,
+  CellInlineContainer,
+  StyledTruncateTooltip,
+};